Split out FHIR function integration tests and fix BlobAcces mocking
